// src/api/openapi/model/tikTakTimeEntryByJob.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserEssential } from './userEssential';


export interface TikTakTimeEntryByJob { 
    hourlyRate?: number | null;
    id: number;
    lastSync: string;
    minutes: number;
    user: UserEssential;
}